// apps/super-app/src/libs/sanity/sanity.types.ts
export declare const internalGroqTypeReferenceTo: unique symbol;
// ...
export interface SanityImageAssetReference {
  _ref: string;
  _type: "reference";
  _weak?: boolean;
  [internalGroqTypeReferenceTo]?: "sanity.imageAsset";
}
// ...
export interface SanityImageCrop {
  _type: "sanity.imageCrop";
  top?: number;
  bottom?: number;
  left?: number;
  right?: number;
}

export interface SanityImageHotspot {
  _type: "sanity.imageHotspot";
  x?: number;
  y?: number;
  height?: number;
  width?: number;
}

export type AiToolGroups = (
  | "ai-tool-image"
  | "ai-tool-text"
  | "ai-tool-model"
  | "ai-tool-code"
  | "ai-tool-doc"
  | "ai-tool-study"
)[];

export interface AiToolPromptSnippet {
  _type: "aiToolPromptSnippet";
  quickTag?: string;
  fullPrompt?: string;
  imagePrompt?: {
    asset?: SanityImageAssetReference;
    media?: unknown;
    hotspot?: SanityImageHotspot;
    crop?: SanityImageCrop;
    _type: "image";
  };
}
// ...
export interface AiGroupConfigLocaleContent {
  _type: "aiGroupConfigLocaleContent";
  title?: string;
  description?: string;
}

export interface AiSeoLocaleContent {
  _type: "aiSeoLocaleContent";
  title?: string;
  description?: string;
  ogImage?: {
    asset?: SanityImageAssetReference;
    media?: unknown;
    hotspot?: SanityImageHotspot;
    crop?: SanityImageCrop;
    _type: "image";
  };
  ogImageAlt?: string;
}
// ...
export interface AiSeo {
  _id: string;
  _type: "aiSeo";
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  name?: string;
  en?: AiSeoLocaleContent;
  zh?: AiSeoLocaleContent;
  th?: AiSeoLocaleContent;
  ar?: AiSeoLocaleContent;
  es?: AiSeoLocaleContent;
  ko?: AiSeoLocaleContent;
  ja?: AiSeoLocaleContent;
  hi?: AiSeoLocaleContent;
}